EPL-5:

+++ src/main/java/com/github/EPLBot/service/TelegramUserServiceImpl.java
@@ -29,7 +29,7 @@ public class TelegramUserServiceImpl implements TelegramUserService {
     }
 
     @Override
-    public Optional<TelegramUser> findByChatId(Integer chatId) {
+    public Optional<TelegramUser> findByChatId(Long chatId) {
         return telegramUserRepository.findById(chatId);
     }
 }
